Remove cast in WindowRecomposer that failed in layoutlib
In WindowRecomposer#createLifecycleAwareWindowRecomposer, we try to cast
context.applicationContext to Application while setting the animation
scale factor. This was throwing a ClassCastException in layoutlib, so as
a workaround we need to wrap the cast into a try/catch block, otherwise
this can lead to problems in Compose Preview, e.g. Interactive Preview
freezing animations.

Fixes: 225964549
Bug: 226158869
Bug: 227155163
Test: N/A. Verified in Compose Preview
Change-Id: I64c9aed564b4fa39ebb95c809c57cacd55d7fcb7