Only invalidateOptionsMenu if action bar exists
The recent addition of the MenuHost APIs have made it so that
invalidateOptionMenu are called much earlier in the Lifecycle of an
AppCompatActivity. This causes the AppCompatDelegateImpl to make an
internal call to ensureSubDecor which causes a crash.

Instead of invalidOptionsMenu in AppCompatDelegateImpl eagerly call
getSupportActionBar, let's call it lazily and instead use
peekSupportActionBar to determine if one is availble. This way we only
make the ensureSubDecor call when it is safe.

Test: NoThemeAppCompatTest
Fixes: 227224798
Change-Id: Ic95bf1c73e15ce34a1d64b7d3ea67b7dff8a507e