Update playground prebuilts
Also updated the script to support mac where the backup file
is mandatory in mac (sed -i ''). We still pass empty to skip
it.

Also needed to change the matcher to match any character instead
of digits. Not sure why but it is not matching the digits on my
mac :)

Bug: n/a
Test: CI