Move outDir set up to settings.gradle
Our previous setup was using subproject {} which violates project
isolation. The new approach is to use getGradle.beforeProject {}
to achieve the same affect and still be compatible with project
isolation.

Also add workaround to code generating projects that require reaching
to buildDir of other projects before these projects are configured.

Test: ./gradlew tasks -> output still in out directory like expected
Change-Id: I3b3a2f5584827bcc74cf662127c77f06ae7c8b55