Default includeFontPadding=true
This CL changes the default includeFontPadding back to true.
Previously, for an alpha release we turned it off in aosp/1992050.

This would break some users when switching to the beta/final release.

Moreover, keeping it experimental and optional for now will enable us
to spend a little more time on clipping, line height, text field
related improvements including defining the default line height
behavior with the functionalities in aosp/2058653.

RelNote: "Turned on default includeFontPadding. It is possible to
turn off the includeFontPadding using TextStyle.platformTextStyle
attribute. In the near future we will change the default behavior
however until that time this allows us to better integrate line height
improvements (aosp/2058653) and solve TextField clipping issues."

Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT

Bug: 171394808
Change-Id: I01423d9a0042a1f3e462236e1fdadb60a20678fc