Move DisposableEffects to be declared after value initialization
The order in which composables are disposed is affected by where the
DisposableEffect is declared. It is best practice to declare the
DisposableEffect immediately after declaring the object the effect is
keyed to.

RelNote: "Composable scopes in NavHost and DialogHost are now disposed
in the expected order, i.e. inner composabled are disposed before outer
composables."
Test: existing tests pass

Change-Id: I157e609fb0609265eef381f729285bf769b56a7c