Update back stack remember lint rule to require key
We have a lint rule to ensure that calls to `getBackStackEntry` within a
composition are always wrapped with a `remember`. Navigation Compose has a
unique scenario where the underlying data may change, but we do not want
to actually animate since you are on the same screen (i.e. singleTop,
canceled animations, etc.). Because we don't want to animate we have to
consider the content the "same" and that choice causes `remember` without
any keys to be insufficient as the content in the remember would not be
disposed.

This change modifies the existing lint rule to ensure a
NavBackStackEntry key is provided to the `remember` function.

Relnote: "The `UnrememberedGetBackStackEntryDetector` lint rule has been
updated to ensure that the `remember` call surrounding the call to
`getBackStackEntry()` also passes in a `NavBackStackEntry` object as a
key."
Bug: 227382831
Test: new Nav lint tests

Change-Id: Ib708126a32c865f758d5967ea177cc9d2c7692d5