Delay the extraction of parameters
Before this change: we extracted all the parameters of all nodes
and cached them on the device.

After this change: when we are in "live mode" stop extracting the
parameters. Instead: when fetching the parameters, we first find
the composable by walking the slot tree again. The composable is
identified via the hash of the anchor.

Note: this change is backwards compatible since:
- an old version of Studio will not specify the new flag:
  delay_parameter_extractions when updateing the settings.
  This code will then continue to extract all parameters as before.

Note: this change is forwards compatible since:
- a newer version of Studio used with an older compose library will
  simply ignore the delay_parameter_extractions. Getting parameters
  will find the parameters in the cache from the node id as before
  since "delayParameterExtractions" will be "false".

Fixes: 229657989
Test: Unit tests added
Change-Id: If1723174182312d6b58213c1e087d67b21240881