ScalingLazyColumn - handle scrollTo/animatedScrollTo when not visible
ScalingLazyColumn/ScalingLazyListState scrollTo/animatedScrollTo are not working correct in some situations when the component is not visible. In order to correctly scroll with AnchorType.Center we need to know the size of the items. To do this scrollTo/animatedScrollTo work by first scrolling the item into the view port, then getting the item size and then doing a second scrollTo/animatedScroll to that will correctly position the item. However if the SLC is not visible then the scrollTo/animatedScrollTo on the underlying LazyColumn does not result in the item being scrolled into view.

This patch does a few related things to correct the behavior
1. Check that the item being scrolled to exists and if not then scroll to the last item.
2. If the scrollTo/animatedScrollTo does not bring the item into view then store the target itemIndex and offset and do the scroll when the item becomes visible again
3. Use the average height of the currently visible list items to compute a more accurate first guess of where the item should be located on the screen
4. Check whether the item is in the correct place before doing the final fine grained adjustment of the item
5. Fix a typo in the picker docs

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 230582961, 230617988
Change-Id: I3cbff51c64501972fea530c66f648d347c431e33