Fix PhotoPicker stack overflow issue
The PickVisualMediaRequest builder uses the top level Kotlin function to
build a request, which uses the builder, causing a stack overflow due to
recursion.

Fixing this error by making the builder call set the mediaType.

RelNote: "Fixed an error in `PickVisualMediaRequest` where attempting to
create a request, either via the Builder or the top level Kotlin
function, would result in a stack overflow."
Bug: 230930287
Test: Added PickVisualMediaRequestTest

Change-Id: Id929508bdd754bcd3b6995d641255f650d445201