[ZSL] Make RingBuffer thread safe and catch exceptions when acquireLatestImage and queueInputImage
The thread safety is necessary to avoid enqueuing images which might be dequeued and closed at the same time.
ImageReader.acquireLatestImage will throw exception when image reader has reached max images, in that case, we skip that image. ImageWriter.queueInputImage will throw exception when image is already closed, in t will throw exception when image is already closed, in that case, we fall back to regular capture.

Bug: b/231161542
Test: Take picture in regular/zsl capture mode on devices
Change-Id: Ia26b7708370f3da3eb4dacb1b55a4be68e97928c