Fix broken LazyListScrollingBenchmark
Some of the lazy lists benchmarks started crashing after we merged aosp/2056467. The reason for it is now we do not dispose the items we kept for reuse straight away. Instead when it happens we just switch active flag backed by a MutableState. This causes recomposition on the next frame where the content will be partially disposed keeping only the LayoutNodes we can reuse. It felt as a good idea to not do this disposing right during the measuring in order to do less work on each scrolling and postponing some of the work which is not really visible for the user till the next frame. But as a sideeffect the benchmarks were broken as they assert that there will be no pending changed scheduled. I fixed it by also running the following recomposition after the toggle so we benchmark the whole work needed for the scroll including this content disposal.

Fixes: 229268397
Test: run LazyListScrollingBenchmark locally
Change-Id: I75fc30233e2caf0504897aa56e8ed8464048c7db