Set the AccessibilityNodeInfo classname for grid to GridView
A11y services identified grids and lists by number of columns/rows. A grid with multiple columns and rows may
be identifed as a list if it fills only one row or column. Ex. A 2 x 2 grid has only one item.
This will be shown as a Collection with 2 col/row with 1 col/row and be identified as a list.  So we guarantee
 it is identifed as a grid by setting a class name.

Relnote: "Ensure grids are treated as grids by a11y services by setting an a11y node info class name."
Test: Manual with support-v7-demos
Bug: 214605973
Change-Id: I128122352a3fc4898cc6ba76d7da8166519bd648