Update the Picker samples
Updating Picker samples in line with discussions we had about how to improve the API for Horologist.
Remember the date/time used for the pickers so that when you reenter them they have the updated value.
Make the pickers close and update the remembered value when the confirm icon is selected
Bug: 233061347
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have made some improvements to the Picker integration test demos."

Change-Id: I63dd9acde26fded332cbbc7dd4b8ad2ff8266086