Fix inner text field position of the OutlinedTextField when label is multiline
When we calculated the height of the OutlinedTextField, we take into
account the half of label's height. However, when we placed the inner
text field, we didn't take into account in both multi and single line
OutlinedTextField. As a result, when label was a multiline, the inner
text would overlap with the label. This CL ensures that we position our
inner text below the label.

Fixes: 233360348
Test: new tests in TextFieldDecorationBoxTest
Change-Id: Ia8b78ab0effb788e1ec880a6bea2365d26dc9e96