Add the collectAsStateWithLifecycle APIs
Relnote: """
Add new [State]Flow.collectAsStateWithLifecycle APIs that collect
from flows and represents its latest value as Compose State in a
lifecycle-aware manner. The flow is collected and the new emission
is set to the State's value when the lifecycle is at least in a
certain Lifecycle.State. When the lifecycle falls below that
Lifecycle.State, the flow collection stops and the State's value
is not updated.
"""
Fixes: b/230557927
Test: CollectAsStateWithLifecycleTests

Change-Id: I1856e689232fedf11c133907a63bc08b690d2ce9