Create an empty test app for intern project
The intern project is a test app that can be used by 3P developers to diagnose device quirks. By using this app, developers should be able to tell if the bug they have is a CameraX bug or a device quirk. This would give us the ability to debug device quirks remotely and save time.

Bug: 233476986
Test: manual test and ./gradlew bOS
Change-Id: I5df7f680f9e27a643331b126d22f63dc01326d5e