Remove limitation to CompactChip that it can only have 9 characters.
This limitation is too strict. In order to prevent chip bleeding off
the edge of the screen when it is at the bottom, we've put extra
padding to the box where chip is in PrimaryLayout for round devices.

Bug: 229136373
Test: Added
Relnote: "We have removed limitation of 9 character on CompactChip. If text is too big to fit into screen in PrimaryLayout, it will be ellipsized."
Change-Id: Id56eca1d31999330657ce8a4ad3d260ab039c16b