Increase the tap target height of the CompactChip
The CompactChip is only 32.dp high and as a result does not meet a11y guidelines which require tap targets to be 48.dp minimum. Add top/bottom padding of 8.dp to ensure that the tap area is at least 48.dp high to met accessibility guidelines.

Bug: 233990645
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have added padding to the CompactChip in order to ensure that its tap target size is at least 48.dp high to meet the Material accessibility guidelines. This might impact any layouts you have using CompactChips as they will be taking up additional space."
Change-Id: I3d57caebc66e5d4c867f2ce19b4d0d9b5ebc9998