Updates to implementation & Javadoc of all fromLayoutElement methods
Adding check into all methods that if passed in parameter is wanted component,
to just return it. Clarify the Javadoc on these methods on how passed in
LayoutElement should be obtained.

Adding more explanation in classes' Javadocs for purpose and usage of
these methods.

Bug: 233948894
Bug: 233212557
Test: Added :)
Relnote: "Javadoc improvements for fromLayoutElement method in Material."
Change-Id: I80f31a7efe8b45dfc8be733d82217c8ab23f8ba0